requirements.txt
monai[all]==1.4.0
nibabel>=5.2
scikit-learn>=1.4
matplotlib>=3.8
tqdm>=4.66
wandb>=0.19
pyyaml>=6.0
scipy>=1.12
numpy>=1.24,<2.0
